requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.18.4
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asn1crypto==1.5.1
asyncpg==0.31.0
attrs==26.1.0
Authlib==1.6.9
blinker==1.9.0
Bottleneck==1.6.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cloudpickle==3.1.2
cryptography==46.0.5
dataclasses-json==0.6.7
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.135.2
filetype==1.2.0
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.46
google-adk==1.27.4
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-cloud-aiplatform==1.143.0
google-cloud-alloydb==0.8.0
google-cloud-alloydb-connector==1.12.1
google-cloud-appengine-logging==1.8.0
google-cloud-audit-log==0.4.0
google-cloud-bigquery==3.40.1
google-cloud-bigquery-storage==2.36.2
google-cloud-bigtable==2.35.0
google-cloud-core==2.5.0
google-cloud-dataplex==2.17.0
google-cloud-discoveryengine==0.13.12
google-cloud-iam==2.21.0
google-cloud-logging==3.14.0
google-cloud-monitoring==2.29.1
google-cloud-pubsub==2.36.0
google-cloud-resource-manager==1.16.0
google-cloud-secret-manager==2.26.0
google-cloud-spanner==3.63.0
google-cloud-speech==2.37.0
google-cloud-storage==3.10.1
google-cloud-trace==1.18.0
google-cloud-vectorsearch==0.8.0
google-crc32c==1.8.0
google-genai==1.68.0
google-resumable-media==2.8.0
googleapis-common-protos==1.73.0
graphviz==0.21
greenlet==3.3.2
grpc-google-iam-v1==0.14.3
grpc-interceptor==0.15.4
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.22
langchain-google-alloydb-pg==0.14.0
langchain-google-vertexai==3.2.2
langchain-postgres==0.0.17
langchain-tests==1.1.5
langchain-text-splitters==1.1.1
langgraph-checkpoint==3.0.1
langsmith==0.7.22
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mcp==1.26.0
mdurl==0.1.2
mmh3==5.2.1
multidict==6.7.1
mypy_extensions==1.1.0
narwhals==2.18.1
numexpr==2.14.1
numpy==2.4.3
opentelemetry-api==1.38.0
opentelemetry-exporter-gcp-logging==1.11.0a0
opentelemetry-exporter-gcp-monitoring==1.11.0a0
opentelemetry-exporter-gcp-trace==1.11.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-resourcedetector-gcp==1.11.0a0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pandas==2.3.3
pg8000==1.31.5
pgvector==0.3.6
pillow==12.1.1
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.6
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg-pool==3.3.0
py-cpuinfo==9.0.0
pyarrow==22.0.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.12.1
pyOpenSSL==26.0.0
pyparsing==3.3.2
pypdf==6.9.2
PyPDF2==3.0.1
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-benchmark==5.2.3
pytest-codspeed==4.3.0
pytest-recording==0.13.4
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.0
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.30.0
scramp==1.4.8
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
SQLAlchemy==2.0.48
sqlalchemy-spanner==1.17.3
sqlparse==0.5.5
sse-starlette==3.3.3
starlette==0.52.1
streamlit==1.55.0
syrupy==5.1.0
tenacity==9.1.4
toml==0.10.2
tornado==6.5.5
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.42.0
validators==0.35.0
vcrpy==8.1.1
vertexai==1.43.0
watchdog==6.0.0
websockets==15.0.1
wrapt==2.1.2
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
zstandard==0.25.0
